Restructure files, fix issue noted in feedback
