+++ aries-site/src/examples/components/tag/CreateTag/AppContainer.js
@@ -1,0 +1,13 @@
+import { Box } from 'grommet';
+
+export const AppContainer = ({ ...rest }) => (
+  <Box
+    gap="large"
+    fill
+    margin="auto"
+    overflow="auto"
+    pad="medium"
+    width={{ max: 'xxlarge' }}
+    {...rest}
+  />
+);
