+++ aries-site/src/examples/components/tag/CreateTag/TagPageHeader.js
@@ -1,0 +1,17 @@
+import { Box, Heading, Text } from 'grommet';
+
+export const TagPageHeader = () => (
+  <Box gap="large" flex={false}>
+    <Box gap="xsmall" flex={false}>
+      <Heading level={2} margin="none">
+        Create and Assign Tags
+      </Heading>
+      <Text size="large">
+        Tags are name-value pairs that can be assigned to resources.
+      </Text>
+    </Box>
+    <Text size="large">
+      Tags will be assigned to <Text weight="bold">1</Text> device.
+    </Text>
+  </Box>
+);
