+++ aries-site/src/examples/components/tag/CreateTag/index.js
@@ -1,0 +1,5 @@
+export * from './AppContainer';
+export * from './CreateTag';
+export * from './TagPageHeader';
+export * from './TagResults';
+export * from './utils';
