+++ aries-site/src/examples/components/tag/index.js
@@ -1,4 +1,4 @@
-export * from './CreateNameValuePair';
+export * from './CreateTag';
 export * from './data';
 export * from './TagBasicAnatomy';
 export * from './TagAnatomy';
